catch-the-ball/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
redis==5.0.1
cryptography==41.0.3
ratelimit==2.2.1 